add processes to get products and purchases
we don't actually need docker for any of this, but I kept in the Dockerfiles anyway. Once the entire app is up and running, I'll probably massively refactor, since there's lots of repeated go code at the moment. Now just need to add in reviews and referrals and we'll be off to the races!